Make it safer

+++ administrator/components/com_fields/helpers/fields.php
@@ -296,8 +296,8 @@ class FieldsHelper
 
 		// Account for case that a submitted form has a multi-value category id field (e.g. a filtering form), just use the first category
 		$assignedCatids = is_array($assignedCatids)
-			? reset($assignedCatids)
-			: $assignedCatids;
+			? (int) reset($assignedCatids)
+			: (int) $assignedCatids;
 
 		if (!$assignedCatids && $formField = $form->getField('catid'))
 		{
